:recycle: 캐싱된 데이터에서 가장 인접한 열차 가져오도록 수정 (#175)

+++ ahachul_backend/src/main/kotlin/backend/team/ahachul_backend/common/client/TrainCongestionClient.kt
@@ -17,7 +17,7 @@ class TrainCongestionClient(
     private val publicDataProperties: PublicDataProperties
 ) {
 
-    fun getCongestions(subwayLine: Int, trainNo: Int): TrainCongestionDto {
+    fun getCongestions(subwayLine: Long, trainNo: Int): TrainCongestionDto {
         val url = "${publicDataProperties.realTimeCongestionUrl}/$subwayLine/$trainNo"
 
         try {
